song-service entry data
Getting the hang on how I want the song service to look like. I created the BPM and Song Key finder logic using a python script and running that on a small flask application.
Also made some changes to the front end, changed the sidebar into a more classic navigation bar. Felt that the sidebar looked a lot like Spotify and didn't fit what I was trying to achieve in the design process.
Going forward I'm going to be looking into creating the final look of the song database schema, creating at least the basic CRUD operations *and maybe some web-socket for the messaging*